Here is what I did today
controllers\ApplicationController.php I created a method that will
eventually GET the issues from the github api and store them in $issues.
Then I will pass $issues (which hopefully is an array of issues) to
manage/manageIssues to display it.

views\manage]manageIssues.php I am testing the flight:render trying to
pass $issues. I get 3 errors...... and for somereason $issues is set to
false. Is this because we don't have an account set up in the tracker?