work on issues
@gofishAOD more stuff on issues. I went ahead and implemented the label
filter we discussed before, since it was newly developed. Also did a few
other things:

- updated the route to use the new controller
- updated the listIssues method with the newer listAllIssues and added
some documentation explaining the parameters
- created an excerpt method for shortening bodies of text
- some cosmetic work on the issues listing page, went with panels
instead of list-group-items, seemed a bit cleaner. Plus we can use the
footer for other things if we need to.

All we need now is the view for individual issues, and we can move onto
creating the handlers for submitting issues, and maybe comments