some cleanup
I'm not overly worried about checking dev issues against access. Right
now it's not possible for a user to even be able to see which issues are
dev issues, so it's highly unlikely that they would be able to view one.

There's also not a mechanism in place to check the existence of an
issue, so we'll have to approach that again later.

I did a bit of cleanup on your recent work. One important thing to
consider when writing code is thinking about the most efficient way to
do what you're doing. Are you repeating code unnecessarily? Is there a
better way? Keeping this in mind will help us in the future when updates
are needed, and we don't have a huge mess on our hands